# swagger_config.yaml
swagger: "2.0"
info:
  title: "Volumx Business API"
  description: "Endpoints for managing businesses and user rights"
  version: "1.0"
schemes:
  - http
  - https
...
paths:
  /api/v1/business/{business_id}:
    get:
      tags:
        - Business
      summary: Get business data by ID
      description: Retrieve details of a specific business by ID
      parameters:
        - name: business_id
          in: path
          required: true
          type: string
          description: ID of the business
      responses:
        "200":
          description: Successful response
      "404":
        description: Business not found

    put:
      tags:
        - Business
      summary: Update business by ID
      description: Update details of a specific business by ID
      parameters:
        - name: business_id
          in: path
          required: true
          type: string
          description: ID of the business
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/definitions/BusinessUpdate"
      responses:
        "200":
          description: Successful response
        "400":
          description: Bad request

    patch:
      tags:
        - Business
      summary: Patch business by ID
      description: Partially update details of a specific business by ID
      parameters:
        - name: business_id
          in: path
          required: true
          type: string
          description: ID of the business
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/definitions/BusinessPatch"
      responses:
        "200":
          description: Successful response
        "400":
          description: Bad request
        "404":
          description: Business not found

    delete:
      tags:
        - Business
      summary: Delete business by ID
      description: Remove a specific business by ID
      parameters:
        - name: business_id
          in: path
          required: true
          type: string
          description: ID of the business
      responses:
        "200":
          description: Successful response
        "400":
          description: Bad request
        "404":
          description: Business not found
  /api/v1/business/:
    get:
      tags:
        - Business
      summary: Get all businesses
      description: Retrieve details of all businesses
      responses:
        "200":
          description: Successful response
      "404":
        description: Businesses not found

    post:
      tags:
        - Business
      summary: Create a new business
      description: Register a new business
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/definitions/BusinessCreate"
      responses:
        "201":
          description: Successful creation
          content:
            application/json:
              schema:
                $ref: "#/definitions/BusinessResponse"
              example:
                message: Business created successfully
                businessId: "12345"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/definitions/Error"
              example:
                error: Invalid data provided
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/definitions/Error"
              example:
                error: User not found
  /api/v1/business/user_rights/{user_id}:
    get:
      tags:
        - Business User Rights
      summary: Get business user rights by User ID
      description: Retrieve user rights for a specific user
      parameters:
        - name: user_id
          in: path
          required: true
          type: string
          description: ID of the user
      responses:
        "200":
          description: Successful response
      "404":
        description: User not found
  /api/v1/business/user_rights:
    get:
      tags:
        - Business User Rights
      summary: Get all business user rights
      description: Retrieve user rights for all businesses
      responses:
        "200":
          description: Successful response
      "404":
        description: Business user rights not found

    post:
      tags:
        - Business User Rights
      summary: Create business user rights
      description: Register user rights for a business
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/definitions/BusinessUserRightsCreate"
      responses:
        "201":
          description: Successful response
        "400":
          description: Bad request
        "404":
          description: Business not found
...
definitions:
...
  BusinessCreate:
    type: object
    properties:
      legalName:
        type: string
      displayName:
        type: string
      # Include other properties
  BusinessUpdate:
    type: object
    properties:
      legalName:
        type: string
      displayName:
        type: string
      # Include other properties
  BusinessPatch:
    type: object
    properties:
      legalName:
        type: string
      displayName:
        type: string
      # Include other properties
  BusinessUserRightsCreate:
    type: object
    properties:
      userId:
        type: string
      businessId:
        type: string
      # Include other properties